Store genome on Polygon in Reveal

+++ contracts/tests/GenomeProviderTest.sol
@@ -9,23 +9,23 @@ contract GenomeProviderTest is GenomeProviderPolygon {
     constructor(uint256 _secret_seed_hash) GenomeProviderPolygon(_secret_seed_hash) {
     }
 
-    function getTraitWeightSum(uint traitId) external view returns (uint) {
+    function getTraitWeightSum(uint traitId) external view returns (uint32) {
         return traits[traitId].sum;
     }
 
-    function getTraitSetupData(uint traitId) external view returns (uint[] memory) {
+    function getTraitSetupData(uint traitId) external view returns (uint8[] memory) {
         return traits[traitId].traitData;
     }
 
-    function getTraitSetupSlots(uint traitId) external view returns (uint[] memory) {
+    function getTraitSetupSlots(uint traitId) external view returns (uint8[] memory) {
         return traits[traitId].slots;
     }
 
-    function getTraitWeight(uint traitId) external view returns (uint[] memory) {
+    function getTraitWeight(uint traitId) external view returns (uint16[] memory) {
         return traits[traitId].weight;
     }
 
-    function getTraitWeightMaxBonus(uint traitId) external view returns (uint[] memory) {
+    function getTraitWeightMaxBonus(uint traitId) external view returns (uint16[] memory) {
         return traits[traitId].weightMaxBonus;
     }
